The game of Scrabble
In the game of Scrabble, players create words to score points, and the number of points is the sum of the point values of each letter in the word. The player scores the most points is the final winner. The program prints Player 1 wins!, Player 2 wins!, or Tie! for respective scenarios.